requirements.txt
streamlit==1.12.2
gsheetsdb==0.1.13.1
gspread_pandas==3.2.2
